# backend/api/build/resources/main/openapi/dopashift-api-v1.yml
openapi: 3.0.3
info:
  title: DopaShift API
  description: |
    REST API for the DopaShift multi-goal life-tracking platform.
    
    ## Versioning Policy
    - All endpoints are prefixed with `/v1/`
    - Previous API versions are supported for 6 months after a new version is released.
    - Breaking changes will only be introduced in new major versions.
    
    ## Authentication
    All endpoints (except /v1/auth/*) require a valid Bearer token in the Authorization header.
    Tokens are obtained via the Keycloak-proxied authentication endpoints.
  version: 1.0.0
  contact:
    name: DopaShift Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dopashift.app
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/goals:
    post:
      tags: [Goals]
      operationId: createGoal
      summary: Create a new goal profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Goal name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Goals]
      operationId: listGoals
      summary: List all goals for the authenticated user
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalResponse'
  /v1/goals/{id}:
    get:
      tags: [Goals]
      operationId: getGoal
      summary: Get goal detail with keywords and stats
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Goals]
      operationId: updateGoal
      summary: Update an existing goal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Goals]
      operationId: deleteGoal
      summary: Delete a goal with dependency resolution
      parameters:
        - $ref: '#/components/parameters/GoalId'
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum: [DELETE_ALL, REASSIGN]
      responses:
        '204':
          description: Goal deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # === Checklist ===
...
  /v1/habits:
    post:
      tags: [Habits]
      operationId: activateHabitTrack
      summary: Activate a 30-day habit track for a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateHabitRequest'
      responses:
        '201':
          description: Habit track activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitTrackResponse'
    get:
      tags: [Habits]
      operationId: listHabitTracks
      parameters:
        - name: goalId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Habit tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HabitTrackResponse'
  /v1/habits/{id}:
    get:
      tags: [Habits]
      operationId: getHabitTrack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Habit track with all checkpoints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitTrackDetailResponse'
  /v1/habits/{id}/checkpoints/{day}:
    put:
      tags: [Habits]
      operationId: markCheckpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: day
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkCheckpointRequest'
      responses:
        '200':
          description: Checkpoint updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointResponse'
  /v1/habits/current:
    get:
      tags: [Habits]
      operationId: getCurrentHabitForOverlay
      summary: Get earliest active habit track (for overlay display)
      responses:
        '200':
          description: Current habit track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitTrackDetailResponse'
        '204':
          description: No active habit tracks

  # === Reminders ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GoalId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, correlationId, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Goal name must be 1-100 characters"
        details:
          type: object
          additionalProperties: true
        correlationId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # === Auth ===
...
    CreateGoalRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        keywords:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 20
    UpdateGoalRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        keywords:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 20
    GoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GoalDetailResponse:
      allOf:
        - $ref: '#/components/schemas/GoalResponse'
        - type: object
          properties:
            checklistItems:
              type: array
              items:
                $ref: '#/components/schemas/ChecklistItemResponse'
            activeHabitTracks:
              type: integer
            completionPercentage:
              type: number
              format: float

    # === Checklist ===
...
    ChecklistItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        text:
          type: string
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # === Todos ===
...
    ActivateHabitRequest:
      type: object
      required: [goalId, checkpoints]
      properties:
        goalId:
          type: string
          format: uuid
        checkpoints:
          type: array
          items:
            type: object
            properties:
              dayNumber:
                type: integer
                minimum: 1
                maximum: 30
              description:
                type: string
                minLength: 1
                maxLength: 200
          minItems: 30
          maxItems: 30
    HabitTrackResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        currentDay:
          type: integer
        isFinished:
          type: boolean
    HabitTrackDetailResponse:
      allOf:
        - $ref: '#/components/schemas/HabitTrackResponse'
        - type: object
          properties:
            checkpoints:
              type: array
              items:
                $ref: '#/components/schemas/CheckpointResponse'
    MarkCheckpointRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [COMPLETED, MISSED]
    CheckpointResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dayNumber:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [PENDING, COMPLETED, MISSED]
        completedAt:
          type: string
          format: date-time

    # === Reminders ===
...